refactor: maintain score
