+++ internal/service/hub/handler/dsl/enforcer/enforcer.go
@@ -2,9 +2,7 @@ package enforcer
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"io"
 	"os"
 
 	"github.com/ethereum/go-ethereum/common"
@@ -14,7 +12,6 @@ import (
 	"github.com/rss3-network/global-indexer/internal/database"
 	"github.com/rss3-network/global-indexer/internal/service/hub/handler/dsl/model"
 	"github.com/rss3-network/global-indexer/schema"
-	"github.com/rss3-network/node/schema/worker/decentralized"
 	"github.com/samber/lo"
 	"go.uber.org/zap"
 )
@@ -54,16 +51,17 @@ func (e *SimpleEnforcer) VerifyResponses(ctx context.Context, responses []*model
 	updatePointsBasedOnIdentity(responses)
 	// update stats struct based on the above results
 	updateStatsWithResults(nodeStatsMap, responses)
+
+	nodeStatsSlice := lo.MapToSlice(nodeStatsMap, func(_ common.Address, stat *schema.Stat) *schema.Stat {
+		return stat
+	})
 	// save stats to the database
-	if err = e.databaseClient.SaveNodeStats(ctx, lo.MapToSlice(nodeStatsMap,
-		func(_ common.Address, stat *schema.Stat) *schema.Stat {
-			return stat
-		})); err != nil {
+	if err = e.databaseClient.SaveNodeStats(ctx, nodeStatsSlice); err != nil {
 		return fmt.Errorf("save Node stats: %w", err)
 	}
 
 	// update the score maintainer
-	e.updateScoreMaintainer(ctx, nodeStatsMap)
+	e.batchUpdateScoreMaintainer(ctx, nodeStatsSlice)
 
 	return nil
 }
@@ -103,222 +101,6 @@ func (e *SimpleEnforcer) VerifyPartialResponses(ctx context.Context, epochID uin
 	e.verifyPartialActivities(ctx, epochID, responses[0], activities.Data, workingNodes)
 }
 
-func (e *SimpleEnforcer) getNodeStatsMap(ctx context.Context, responses []*model.DataResponse) (map[common.Address]*schema.Stat, error) {
-	stats, err := e.databaseClient.FindNodeStats(ctx, &schema.StatQuery{
-		Addresses: lo.Map(responses, func(response *model.DataResponse, _ int) common.Address {
-			return response.Address
-		}),
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	return lo.SliceToMap(stats, func(stat *schema.Stat) (common.Address, *schema.Stat) {
-		return stat.Address, stat
-	}), nil
-}
-
-func updateStatsWithResults(statsMap map[common.Address]*schema.Stat, responses []*model.DataResponse) {
-	for _, response := range responses {
-		if stat, exists := statsMap[response.Address]; exists {
-			stat.TotalRequest += int64(response.ValidPoint)
-			stat.EpochRequest += int64(response.ValidPoint)
-			stat.EpochInvalidRequest += int64(response.InvalidPoint)
-		}
-	}
-}
-
-func (e *SimpleEnforcer) updateScoreMaintainer(ctx context.Context, nodeStatsMap map[common.Address]*schema.Stat) {
-	for _, stat := range nodeStatsMap {
-		_ = CalculateReliabilityScore(stat)
-
-		nodeCache := &model.NodeEndpointCache{
-			Address:      stat.Address.String(),
-			Score:        stat.Score,
-			Endpoint:     stat.Endpoint,
-			InvalidCount: stat.EpochInvalidRequest,
-		}
-
-		if err := e.fullNodeScoreMaintainer.addOrUpdateScore(ctx, model.FullNodeCacheKey, nodeCache); err != nil {
-			zap.L().Error("failed to update full node score", zap.Error(err))
-		}
-
-		if err := e.rssNodeScoreMaintainer.addOrUpdateScore(ctx, model.RssNodeCacheKey, nodeCache); err != nil {
-			zap.L().Error("failed to update rss node score", zap.Error(err))
-		}
-	}
-}
-
-// verifyPartialActivities filter Activity based on the platform to perform a partial verification.
-func (e *SimpleEnforcer) verifyPartialActivities(ctx context.Context, epochID uint64, validResponse *model.DataResponse, activities []*model.Activity, workingNodes []common.Address) {
-	// platformMap is used to store the platform that has been verified
-	platformMap := make(map[string]struct{}, model.RequiredVerificationCount)
-	// statMap is used to store the stats that have been verified
-	statMap := make(map[string]struct{})
-
-	nodeInvalidResponse := &schema.NodeInvalidResponse{
-		EpochID:       epochID,
-		VerifierNodes: []common.Address{validResponse.Address},
-	}
-
-	for _, activity := range activities {
-		// This usually indicates that the activity belongs to the fallback worker.
-		// We cannot determine whether this activity belongs to a readable worker，
-		// therefore it is skipped.
-		if len(activity.Platform) == 0 {
-			continue
-		}
-
-		// Find stats that related to the platform
-		stats, err := e.findStatsByPlatform(ctx, activity, workingNodes)
-
-		if err != nil {
-			zap.L().Error("failed to verify platform", zap.Error(err))
-
-			continue
-		}
-
-		if len(stats) == 0 {
-			zap.L().Warn("no stats match the platform")
-
-			continue
-		}
-
-		e.verifyActivityByStats(ctx, activity, stats, statMap, platformMap, nodeInvalidResponse)
-
-		// If the platform count reaches the RequiredVerificationCount, exit the verification loop.
-		if _, exists := platformMap[activity.Platform]; !exists {
-			if len(platformMap) == model.RequiredVerificationCount {
-				break
-			}
-		}
-	}
-}
-
-// findStatsByPlatform finds the required stats based on the platform.
-func (e *SimpleEnforcer) findStatsByPlatform(ctx context.Context, activity *model.Activity, workingNodes []common.Address) ([]*schema.Stat, error) {
-	pid, err := decentralized.PlatformString(activity.Platform)
-	if err != nil {
-		return nil, err
-	}
-
-	workers := model.PlatformToWorkersMap[pid.String()]
-
-	indexers, err := e.databaseClient.FindNodeWorkers(ctx, &schema.WorkerQuery{
-		Networks: []string{activity.Network},
-		Names:    workers,
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	nodeAddresses := excludeWorkingNodes(indexers, workingNodes)
-
-	if len(nodeAddresses) == 0 {
-		return nil, nil
-	}
-
-	stats, err := e.databaseClient.FindNodeStats(ctx, &schema.StatQuery{
-		Addresses:    nodeAddresses,
-		ValidRequest: lo.ToPtr(model.DemotionCountBeforeSlashing),
-		PointsOrder:  lo.ToPtr("DESC"),
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	return stats, nil
-}
-
-// excludeWorkingNodes excludes the working Nodes from the indexers.
-func excludeWorkingNodes(indexers []*schema.Worker, workingNodes []common.Address) []common.Address {
-	nodeAddresses := lo.Map(indexers, func(indexer *schema.Worker, _ int) common.Address {
-		return indexer.Address
-	})
-
-	// filter out the working nodes
-	return lo.Filter(nodeAddresses, func(item common.Address, _ int) bool {
-		return !lo.Contains(workingNodes, item)
-	})
-}
-
-// verifyActivityByStats verifies the activity based on stats nodes that meet specific criteria.
-func (e *SimpleEnforcer) verifyActivityByStats(ctx context.Context, activity *model.Activity, stats []*schema.Stat, statMap, platformMap map[string]struct{}, nodeInvalidResponse *schema.NodeInvalidResponse) {
-	for _, stat := range stats {
-		if _, exists := statMap[stat.Address.String()]; !exists {
-			statMap[stat.Address.String()] = struct{}{}
-
-			activityFetched, err := e.fetchActivityByTxID(ctx, stat.Endpoint, activity.ID)
-
-			if err != nil || activityFetched.Data == nil || !isActivityIdentical(activity, activityFetched.Data) {
-				stat.EpochInvalidRequest += invalidPointUnit
-
-				nodeInvalidResponse.Type = lo.Ternary(err != nil, schema.NodeInvalidResponseTypeError, schema.NodeInvalidResponseTypeInconsistent)
-				nodeInvalidResponse.Response = generateInvalidResponse(err, activityFetched)
-			} else {
-				stat.TotalRequest++
-				stat.EpochRequest += validPointUnit
-			}
-
-			// If the request is invalid, save the invalid response to the database.
-			if stat.EpochInvalidRequest > 0 {
-				nodeInvalidResponse.Node = stat.Address
-				nodeInvalidResponse.Request = stat.Endpoint + "/decentralized/tx/" + activity.ID
-
-				validData, _ := json.Marshal(activity)
-				nodeInvalidResponse.VerifierResponse = validData
-
-				if err = e.databaseClient.SaveNodeInvalidResponses(ctx, []*schema.NodeInvalidResponse{nodeInvalidResponse}); err != nil {
-					zap.L().Error("save node invalid response", zap.Error(err))
-				}
-			}
-
-			platformMap[activity.Platform] = struct{}{}
-
-			if err = e.databaseClient.SaveNodeStat(ctx, stat); err != nil {
-				zap.L().Warn("[verifyStat] failed to save node stat", zap.Error(err))
-			}
-
-			break
-		}
-	}
-}
-
-func generateInvalidResponse(err error, activity *model.ActivityResponse) json.RawMessage {
-	if err != nil {
-		return json.RawMessage(err.Error())
-	}
-
-	rawData, _ := json.Marshal(activity.Data)
-
-	return rawData
-}
-
-// fetchActivityByTxID fetches the activity by txID from a Node.
-func (e *SimpleEnforcer) fetchActivityByTxID(ctx context.Context, nodeEndpoint, txID string) (*model.ActivityResponse, error) {
-	fullURL := nodeEndpoint + "/decentralized/tx/" + txID
-
-	body, err := e.httpClient.Fetch(ctx, fullURL)
-	if err != nil {
-		return nil, err
-	}
-
-	data, err := io.ReadAll(body)
-	if err != nil {
-		return nil, err
-	}
-
-	activity := &model.ActivityResponse{}
-	if isDataValid(data, activity) {
-		return activity, nil
-	}
-
-	return nil, fmt.Errorf("invalid data")
-}
-
 // MaintainReliabilityScore maintains the Reliability Score σ for all Nodes.
 // σ is used to determine the probability of a Node receiving a request on DSL.
 func (e *SimpleEnforcer) MaintainReliabilityScore(ctx context.Context) error {
@@ -337,7 +119,9 @@ func (e *SimpleEnforcer) MaintainReliabilityScore(ctx context.Context) error {
 	return nil
 }
 
-func (e *SimpleEnforcer) MaintainNewEpochData(ctx context.Context, epoch int64) error {
+// MaintainEpochData maintains the data for the new epoch.
+// The data includes the range of data that all nodes can support in a new epoch.
+func (e *SimpleEnforcer) MaintainEpochData(ctx context.Context, epoch int64) error {
 	stats, err := e.getAllNodeStats(ctx)
 	if err != nil {
 		return err
@@ -381,6 +165,31 @@ func (e *SimpleEnforcer) RetrieveQualifiedNodes(ctx context.Context, key string)
 	}
 
 	return nodesCache, err
+}
+
+func NewSimpleEnforcer(ctx context.Context, databaseClient database.Client, cacheClient cache.Client, stakingContract *l2.Staking, httpClient httputil.Client, initScoreMaintainer, initWorkerMap bool) (*SimpleEnforcer, error) {
+	enforcer := &SimpleEnforcer{
+		databaseClient:  databaseClient,
+		cacheClient:     cacheClient,
+		stakingContract: stakingContract,
+		httpClient:      httpClient,
+	}
+
+	if initWorkerMap {
+		if err := enforcer.initWorkerMap(ctx); err != nil {
+			return nil, err
+		}
+	}
+
+	if initScoreMaintainer {
+		if err := enforcer.initScoreMaintainers(ctx); err != nil {
+			return nil, err
+		}
+
+		subscribeNodeCacheUpdate(ctx, cacheClient, databaseClient, enforcer.fullNodeScoreMaintainer, enforcer.rssNodeScoreMaintainer)
+	}
+
+	return enforcer, nil
 }
 
 // subscribeNodeCacheUpdate subscribes to updates of the 'epoch' key.
@@ -447,29 +256,4 @@ func (e *SimpleEnforcer) initScoreMaintainer(ctx context.Context, nodeType strin
 	}
 
 	return newScoreMaintainer(ctx, nodeType, nodes, e.cacheClient)
-}
-
-func NewSimpleEnforcer(ctx context.Context, databaseClient database.Client, cacheClient cache.Client, stakingContract *l2.Staking, httpClient httputil.Client, initScoreMaintainer, initWorkerMap bool) (*SimpleEnforcer, error) {
-	enforcer := &SimpleEnforcer{
-		databaseClient:  databaseClient,
-		cacheClient:     cacheClient,
-		stakingContract: stakingContract,
-		httpClient:      httpClient,
-	}
-
-	if initScoreMaintainer {
-		if err := enforcer.initScoreMaintainers(ctx); err != nil {
-			return nil, err
-		}
-
-		subscribeNodeCacheUpdate(ctx, cacheClient, databaseClient, enforcer.fullNodeScoreMaintainer, enforcer.rssNodeScoreMaintainer)
-	}
-
-	if initWorkerMap {
-		if err := enforcer.initWorkerMap(ctx); err != nil {
-			return nil, err
-		}
-	}
-
-	return enforcer, nil
 }
